The meeting concern gets its own file, and the mark gets its gate
Adding the readiness flag pushed three files past the 500-line cap, which is
the ceiling asking how much a reader must hold at once rather than a formatting
rule. The meeting concern is what they had in common, so it moves out whole:
how a booked meeting becomes an item and how that item is ranked read the same
two facts, and splitting them apart would let the hint and its reader drift.

The mark itself now has the test its comment claimed. A second spelling would
not fail a build — the writer would stamp a word the reader never matches, and
the reason would quietly stop appearing on rows that earned it — so the gate
that fails when the two disagree is what makes the claim true rather than
merely stated.
