+++ backend/internal/compose/attention/render.go
@@ -374,52 +374,6 @@ func lapsedItem(quiet QuietRelationship) crmcontracts.AttentionItem {
 	}
 }
 
-// meetingItem renders one appointment still ahead today.
-//
-// The subject IS the sentence a rep recognises — somebody wrote it when the
-// meeting was booked — so it travels as the title, and the start time as
-// `due_at` because that is what a reader is racing. No `overdue` flag: a
-// meeting that has not started cannot be late, and the lane only carries the
-// ones still ahead.
-//
-// It offers NO verb. The pre-meeting brief is its own surface with its own eight
-// cited sections, and a queue that tried to summarise it here would be a second
-// answer to "prepare me for this". `open` is withheld for a narrower reason than
-// it once was: this card's subject is the ACTIVITY, which is a timeline entry
-// rather than a record with a page of its own, so the verb would advertise a
-// destination that does not exist. The two cards whose subject IS a record —
-// the quiet deal and the promise — now send it.
-func meetingItem(meeting Meeting) crmcontracts.AttentionItem {
-	subject := meeting.Subject
-	starts := meeting.StartsAt
-	item := crmcontracts.AttentionItem{
-		Id:      meeting.ID.String(),
-		Source:  crmcontracts.AttentionItemSource("meeting"),
-		Title:   &subject,
-		Subject: subjectOf("activity", meeting.ID),
-		DueAt:   &starts,
-		Actions: []crmcontracts.AttentionItemActions{},
-	}
-	// `kind` is the producer's own sub-type, for the icon and the label and
-	// never for authority — which is exactly what "nobody has written anything
-	// down for this yet" is. Carried here rather than as a field of its own so
-	// the contract gains no vocabulary for a display hint.
-	//
-	// Set only when the answer is KNOWN. A meeting whose content this reader may
-	// not read arrives with an empty body for a reason that is not preparation,
-	// and `meetingPrep` refuses to guess; an absent kind is that refusal
-	// reaching the page, where it draws nothing.
-	if meeting.PrepKnown && meeting.NeedsPrep {
-		kind := meetingKindUnprepared
-		item.Kind = &kind
-	}
-	return item
-}
-
-// meetingKindUnprepared marks a meeting with nothing written down for it.
-// Spelled once here and read by classifyMeeting, so the two cannot drift.
-const meetingKindUnprepared = "unprepared"
-
 // receiptItem renders one thing the system did on its own.
 //
 // It offers no decision: a receipt reports a finished act, and asking the reader
